+++ ktorfit-lib/src/jvmTest/kotlin/BuilderTest.kt
@@ -1,4 +1,5 @@
 import de.jensklingenberg.ktorfit.Ktorfit
+import de.jensklingenberg.ktorfit.Strings.Companion.EXPECTED_URL_SCHEME
 import de.jensklingenberg.ktorfit.internal.InternalKtorfitApi
 import de.jensklingenberg.ktorfit.internal.KtorfitClient
 import de.jensklingenberg.ktorfit.internal.RequestData
@@ -10,7 +11,36 @@ import org.junit.Assert
 import org.junit.Test
 
 @OptIn(InternalKtorfitApi::class)
-class ClientTest {
+class BuilderTest {
+
+    @Test
+    fun whenBaseUrlNotEndingWithSlashThrowError() {
+        try {
+            val ktorfit = Ktorfit.Builder().baseUrl("http://www.example.com").build()
+        } catch (illegal: IllegalStateException) {
+            assert(illegal.message == "Base URL needs to end with /")
+        }
+    }
+
+    @Test
+    fun whenBaseUrlIsEmptyThrowError() {
+        try {
+            val ktorfit = Ktorfit.Builder().baseUrl("").build()
+        } catch (illegal: IllegalStateException) {
+            assert(illegal.message == "Base URL required")
+        }
+
+    }
+
+    @Test
+    fun whenBaseUrlDontStartsWithHttpThrowError() {
+        try {
+            val ktorfit = Ktorfit.Builder().baseUrl("http://www.example.com/").build()
+        } catch (illegal: IllegalStateException) {
+            assert(illegal.message == EXPECTED_URL_SCHEME)
+        }
+
+    }
 
     @Test
     fun checkIfCorrectHttpMethodIsSet() {
@@ -23,6 +53,7 @@ class ClientTest {
             override fun getRequestData(data: HttpRequestData) {
                 Assert.assertEquals(expectedHTTPMethod, data.method.value)
                 Assert.assertEquals(expectedUrl, data.url.toString())
+
             }
         }
 
@@ -33,7 +64,35 @@ class ClientTest {
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
+
+
     }
+
+    @Test
+    fun checkIfBaseUrlIsSetWhenUrlCheckIsDisabled() {
+        val testBaseUrl = "http://www.example.com"
+        val testRelativeUrl = "posts"
+        val expectedUrl = testBaseUrl + testRelativeUrl
+        val expectedHTTPMethod = "GET"
+
+        val engine = object : TestEngine() {
+            override fun getRequestData(data: HttpRequestData) {
+                Assert.assertEquals(expectedHTTPMethod, data.method.value)
+                Assert.assertEquals(expectedUrl, data.url.toString())
+            }
+        }
+
+        val ktorfit = Ktorfit.Builder().baseUrl(testBaseUrl, false).httpClient(HttpClient(engine)).build()
+        runBlocking {
+            val requestData = RequestData(
+                method = expectedHTTPMethod, relativeUrl = testRelativeUrl, returnTypeData = TypeData("kotlin.String")
+            )
+            KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
+        }
+
+
+    }
+
 
     @Test
     fun whenUrlValueContainsBaseUrl_ThenRemoveBaseUrl() {
@@ -80,6 +139,9 @@ class ClientTest {
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
+
+
     }
+
 }